remove number_of_miners attribute
I've realized that it makes no sense to track the number of miners because the result can be quite ambiguous. Does the number 32 tell you that there are 32 rigs or the first 32 slots are reserved? Therefore, I'll track the status of the miners simply with True, False or None and save the number of racks in the frontend.